Setup for Light Controller Page
Made Lights Folder, changed VFMS to add light button and page location, as well as copy the VFMS index.php to the Light index.php as a base. Commented out sections that are un-needed and added a variable known as slev for determining and displaying the security level of the user (admin, sysadmin, user).